MB-29879: Allow sub-second values for defragmenter_interval
The defragmenter_interval (sleep time) is currently an integer value,
meaning the minimum sleep time (and to not run constantly) is
1s. Given the defragmenter typically runs for only short durations
(~25ms), this is limiting as at most you can have the defragmenter
running for 2.5% of a NonIO thread.

Change the type of defragmenter_interval to double; allowing
fractional sleep times.

Change-Id: I768f948e315a1d4993b0c120e3bd65af86d646cb
Reviewed-on: http://review.couchbase.org/95042
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>