DCP: Fetch priority from underlying connection
As part of the investigation for MB-29973, it was found that the
'priority' field maintained by DcpProducer can be inaccurate, as it's
just a copy of the priority last time that class set it. If other
parts of the code change the connection's priority then it becomes
incorrect.

Avoid this by requesting the priority from the underlying Connection
object (via the a new function in server cookie API) instead of
caching our own.

Change-Id: I52d938f604633815e6f0682639d7ee26790c0598
Reviewed-on: http://review.couchbase.org/95248
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>