MB-30019: Always close replica-checkpoint on memory-snapshot-end
Currently, when a Consumer receives the shapshot-end mutation of a
memory-snapshot we close the current open checkpoint only if
(mem_used > high_watermark).
As seen in MB-30019, we may end up with keeping big checkpoints in
memory when a Consumer has received the latest snapshot.
We this patch we unconditionally close a open replica-checkpoint when
the Consumer receives the snapshot-end mutation for a memory-sanpshot.

Change-Id: I9ed5388dd8f5bbe0b51113ca139049d1825c4790
Reviewed-on: http://review.couchbase.org/95260
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>