MB-29861: Generate a delete time for a deletion during backfill
Deletes need to have delete time set in the expiry field, so that
during compaction tombstones will be purged after a specified
purge interval. But delete times were only being generated in
the checkpoint manager and were not generated for deletes that
were received during a disk backfill, thus causing delete time
to be set as 0. Fix this by also generating the delete time
for those backfill deletes.

Change-Id: Ie835e75e766ee7ac03862d1b13072ff3cbe9e1f9
Reviewed-on: http://review.couchbase.org/95204
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>