MB-30015: Check validity of exptime before checking with purge_before_ts
The compactor checks if the exptime field of the tombstone is less
than purge_before_ts before deciding to purge the item. But the
exptime field needs to have a valid delete time value in order to
do be considered. In case if a tombstones ends up being zero, it
might never get purged which is still a better scenario compared
to a tombstone getting purged much earlier than it should.

Add a configuration parameter retain_erroneous_tombstones to indicate
whether erroneous tombstones need to be retained or not. Add support
for it to be modified at runtime using cbepctl.

Change-Id: I04d22f4dd8ccb94919736c84ea2314eb1212abeb
Reviewed-on: http://review.couchbase.org/95341
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>