MB-30097: Use valueSize() instead of using size directly
The higher order bit of the size of a Blob will be set if the value if
uncompressible. The valueSize() function will clear the bit and return
the actual size of the value. The valueSize() function needs to be used
in the copy contructor of the Blob to use a memcpy instead of using
the size variable directly.

Change-Id: I7745190e83c1ee97c4fdefc3ac06473627159549
Reviewed-on: http://review.couchbase.org/95566
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>