MB-30017: Add BackfillManager::buffer.nextReadSize to stats
The BackfillManager::buffer structure is used to track the memory used
by all backfills of a given connection.
When the buffer is full, the 'nextReadSize' is used to store the size
(in bytes) of the item that we are trying to push to the ready queue of
a DCP Producer.
I add the new 'backfill_buffer_next_read_size' in stats, as it will be
useful for diagnostic.

Change-Id: I930c8f6588d47c5c255d9e32624b847196c2d547
Reviewed-on: http://review.couchbase.org/95617
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>