MB-30017: Add the DcpConsumer buffered ready queue to stats
When a DcpConsumer receives a mutation, it may process the mutation or
buffer it for processing later (e.g, in the case where there is no
available memory when the mutation is received; this status is caught
by the ReplicationThrottle).
For those buffered items, the DcpConsumer keeps the list of VBuckets
in a dedicated 'buffered ready queue'. We already have stats providing
the content of the non-buffered 'ready queue', I add the following new
stats for the buffered case:
    - dcp_buffered_ready_queue_contents
    - dcp_buffered_ready_queue_size
    - dcp_buffered_ready_queue_map_contents
    - dcp_buffered_ready_queue_map_size
Also, I add the new 'processor_notification' stat. That flags determines
the status of the DcpConsumerTask (which is the task responsible for
processing the DCP buffered items).

Change-Id: I34cc27e6e8e0787a1b0287ca250f72e3610f7768
Reviewed-on: http://review.couchbase.org/95665
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>