MB-30022: Don't run all tests through IPv6
The only difference in the memcached source between IPv4 and
IPv6 is when the server socket is created. As a result of
that there isn't much value of running all of unit tests
which tests various logic in memcached on both IPv4 and IPv6.

Create an new test suite to run a basic test on IPv6 just
to see that the server socket is properly coreated, and
that the connection works as expected once set up (by
executing a few number of commands)

Change-Id: I9602dca66db76cbadab33ac3eb31f71b5f04e43a
Reviewed-on: http://review.couchbase.org/95667
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>