Create script to strip results from perfsuite xml files
We will use this to strip the P95 and P99 results from CBNT in order
to reduce the number of test cases we have, and as they do not add much
value which the averages do not already contain. This should hopefully
speed up the CBNT import process and the general speed of accessing
results from the web interface. Further patches will be created to
reduce the overall number of tests which are run.

Usage: python cbnt_perfsuite_strip_results.py -d . -p output -i '.pct99' -i '.pct95'
Will strip any results with .pct99 or .pct95 in their name

Change-Id: I21316962b5a35d07f0b98bff851d94f0da922b59
Reviewed-on: http://review.couchbase.org/95671
Tested-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>