ep-engine: Set max_size default to 100MB
The max_size configuration parameter (which specifies the bucket
quota) previously had a default value of 0 bytes in
configuration.json. This default value is irrelevent in a normal
production configuration, as ns_server will always override it with
the actual bucket quota value when creating an ep-engine instance.

However, in unit tests this value (0) is used unless otherwise
specified. This can cause proeblems in testing:

* in ep_engine.cc, a max_size of zero is mapped to unlimited (which
  could result in runaway tests consuming all resources on the
  system).

* Parts of the system derive their own limits from the overall bucket
  quota (e.g. watermarks; storage cache sizes); and using a limit of 0
  can result in unexpected derived values.

To attempt to make tests more robust / constrained; change the default
value of max_size in configuration.json to 100MB - which is the
minimum bucket size that Couchbase Server allows. Tests can still
override this if they desire - and this patch updates a number which
/do/ require > 100MB.

Change-Id: Ie4a1e0b5e7a3d616201f99fb49abfcd8793f2db2
Reviewed-on: http://review.couchbase.org/95677
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>