MB-30137: Decrement total on-disk items for deletes for full eviction
In in the case of full eviction, the total number of on-disk items
is the source of truth about the total number of items in the
database. Consider the following scenario for a key K:

(i)  store(K, V1): onDiskItems - 1 after persistence
(ii) delete(K): onDiskItems - 1 (before persistence)
(iii) Before persistence of (ii) store(K, V2): onDiskItems - 2

Persistence Callback of (ii) will notice that the rev seqno has
changed because of (iii) and hence will not decrement the
on-disk items. Solution is to decrement the on-disk items on
persistence of a deleted item even if the rev seqno has been
updated

Change-Id: I5ead12e7bdc9d2c6042fae561d1f1537c840de82
Reviewed-on: http://review.couchbase.org/95796
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>