MB-30189: ConnHandler::addStats: use correct cookie for conn priority
As part of http://review.couchbase.org/95248 a ConnHandler's priority
was included in the connection stats; this priority was fetched from
the underlying Connection via it's cookie.

However the wrong cookie was used - the cookie passed to ADD_STAT is
not necessarily a memcached connection cookie; indeed in the case of
KVBucket::snapshotStats() is is an instance of snapshot_stats_t.

As a result, this triggers a crash under AddressSanitizer as we treat
a snapshot_stats_t object as a memcached connection cookie and call
cookie_get_priority on it - this in turn attempts to dereference as a
Cookie object and reads garbage data.

Solution is to use the `cookie` member variable recored when the
ConnHandler object was created; which /is/ the actual memcached
cookie.

Change-Id: Id81ffb748d5c1081572fb9fe3554de1f572c10e5
Reviewed-on: http://review.couchbase.org/95871
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>