MB-30190: Check if Blob is present before calculating uncompressedLength
As identified by UndefinedSanitizer; if we attempt to eject an item
which is compressed then during the HashTable statistics update we
unconditionally dereference the value pointer; which is now null:

    if (mcbp::datatype::is_snappy(v.getDatatype())) {
        size_t uncompressed_length = cb::compression::get_uncompressed_length(
                cb::compression::Algorithm::Snappy,
                {v.getValue()->getData(), v.valuelen()});
                             ^^
                             dereference of potentially null `v`

In the case of UBSan this reports an error. Without UBSan (on
Linux/GCC 7.3 at least) the compiler generates an address offset
calculation without reading memory so getData() returns a bogus
close-to-zero address.

Fix by checking for null before dereferencing the pointer; to simplify
the code the size calculation is moved into StoredValue.

Also expand the HashTable unit tests to catch this - prevously this
was only identified in the larger ep_testsuite tests.

Change-Id: I56dc0cf18a6aeaa547ca55558dd65e92df2f90f0
Reviewed-on: http://review.couchbase.org/95886
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>