MB-30168: Provide error code on BIO_read and BIO_write failures
From https://linux.die.net/man/3/bio_should_retry:

'If BIO_should_retry() returns false then the precise "error condition"
depends on the BIO type that caused it and the return code of the BIO
operation. For example if a call to BIO_read() on a socket BIO returns 0
and BIO_should_retry() is false then the cause will be that the
connection closed.
A similar condition on a file BIO will mean that it has reached EOF'

Change-Id: I468074cfa050f0319df644ed5b4dbfb7abfeeb8f
Reviewed-on: http://review.couchbase.org/95920
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>