MB-30149: Convert EngineIface::initialize to member function
Change EngineIface::initialize from being a function pointer to a pure
virtual member function. This is the first such function to be
converted over; so requires updates to all the different engines to
ensure they construct themselves as subclass of EngineIface.

Note at this point ENGINE_HANDLE and ENGINE_HANDLE_V1 are
interchangable; the various places where we pass 'h' and 'h1' are
redundant as they are the same thing now.

Change-Id: Ibf5db7007cf9e639c7af0f03571f17c4a6acfcd6
Reviewed-on: http://review.couchbase.org/96005
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>