MB-30270: Add protocol support for revoking user permission [1/2]
Add the "on the wire" spec for revoking user permissions to
allow ns_server to implement their side of notifying memcached
about stale entries.

As the internal of the RBAC database isn't specified yet
we cannot implement the logic of this command yet. We could
return a different error code, but that would just make
integration harder. Given that memcached only supports
internal users for now, we'll simply return "success" (and
continue to use the provided rbac file)

Change-Id: I165989bf1f53e67926e425df3f1ef8681ba18483
Reviewed-on: http://review.couchbase.org/96151
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>