MB-30276: Initialise logger with correct log level
The verbosity specified in the memcached.json config file is not
correctly passed to the logger - the logger is always initialised with
level "Info".

The problem is related to the order of construction in
memcached_main():

1. The config file is parsed at memcached.cc:2263, at which point we
   notify all registered callbacks to the change in verbosity
   (including spdlogger).

2. The logger is initialised later at line 2283; where it uses the
   default log level (of info).

As such, the log level from the config file is ignored.

Fix by populating logger::Config::log_level from Settings.verbosity;
so if that has previously been changed then it is correctly populated.

As such, the log level from the config file is ignored.

Change-Id: I428e0dd608b1ea5ad012429aaef69b3ff2f72674
Reviewed-on: http://review.couchbase.org/96168
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>