MB-30149: Convert dcp_interface to base class
Move the engine API's dcp_interface from an EngineIface member
variable (EngineIface::dcp) to a base class which engines implementing
DCP inherit from.

This initial patch doesn't change the definition of dcp_interface
(other than renaming it to DcpIface); it simply updates all uses of
the interface to be treating the engine as a polymorphic type -
i.e. dynamic_cast<> the engine to DcpIface.

Subsequent patches will actually change the members of DcpIface to be
virtual methods.

Change-Id: I4fd85edccf6cf46343f775c48de3722c600a6a7f
Reviewed-on: http://review.couchbase.org/96551
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>