MB-30301: Enable SO_LINGER for unit-test client sockets
Enable SO_LINGER with zero timeout for client connections in unit tests.

This changes the behaviour of close() - any unsent data will be
discarded, and the connection will be immediately closed with a RST,
and is immediately destroyed.  This has the advantage that the socket
doesn't enter TIME_WAIT; and hence doesn't consume an emphemeral port
until it times out (default 60s).

By using LINGER we (hopefully!) avoid issues in CV jobs where
ephemeral ports are exhausted and hence tests intermittently fail. One
minor downside the RST triggers a warning in the server side logs:
'read error: Connection reset by peer'.

Note that this isn't enabled for SSL sockets, which don't appear to be
happy with having the underlying socket closed immediately; I suspect
due to the additional out-of-band messages SSL may send/recv in
addition to normal traffic.

Change-Id: Iaa467ec6041ac2c7c2fffc1347bb8469f0be3303
Reviewed-on: http://review.couchbase.org/96782
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>