Fix readyQ data race in StreamTest.BackfillOnly
At http://review.couchbase.org/#/c/96617 I introduced a check on the
size of the Stream::readyQ to wait for the backfill task to complete.

The backfill task pushes items to the Stream::readyQ. So, it is
necessary to acquire a lock to the streamMutex when accessing
the readyQ while the backfill task is still executing.

Change-Id: Iae5d8c8416f979bf224102e4b444f8d8096f26f8
Reviewed-on: http://review.couchbase.org/96783
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>