ep_testsuite: check compression mode via engine API
Change the implementation of is{,Passive,Ative}CompressionEnabled() to
use the direct EngineIface method getCompressionMode(); instead of
making a stats call for the top-level group and extracting
ep_compression_mode from it.

This is a significantly faster implmementation, particulary under the
sanitizers - calling get_str_stat() requires a signficiant amount of
string processing.

For example, when running under TSan this reduces the runtime of
ep_testsuite test 'kill -9 bucket' by an order of magnitude - from
18.9s to 1.79s.

Change-Id: I9274acc0a7e139124b645152a2aa48ae836b860f
Reviewed-on: http://review.couchbase.org/96854
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>