test warmup oom: skip if no memory_tracker
The test 'warmup oom' cannot be reliably run without an accurate
memory tracker; as it relies on that to check the
watermark. Furthermore, it takes a long time (12s) to run under
ThreadSanitizer (which doesn't have memory tracking).

As such, disable if memory tracker isn't available. Also only run it
in the full_eviction config given that's a requirement.

Change-Id: Id94f075f5af48cf78c4c253496f096645b22b604
Reviewed-on: http://review.couchbase.org/96865
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>