MB-30518: Collections by-ID conversion
Change collections from a key-prefix (by-name) implementation
to a ID based implementation. With by-name clients had to encode
keys in a way that included the collection name "users:01aa" with
by-ID the client encodes the ID of the collection a key belongs
to and the protocol command now says "set key=01aa collection=22".
A collection 22 maps to a name e.g. "users".

To achieve the switch we first update DocNamespace to be a uint32_t,
no longer a byte sized enum, we also introduce CollectionID as the
formal name for identifying keys as part of the DocKey abstraction.
DocNamespace remains to reduce the patch size, but can be removed fully
soon. CollectionID names two values, 0 for the DefaultCollection and 1
for internal use (where we will write 'collection-create' events).

The Collection manifest has always assigned uid's to collections

  Collection-Entry {"name" : "name", "uid" : "uid-value"}

This existing UID is the identifier of the collection.

The Collections::Manifest is the only object to store the uid/name
mapping, once a Manifest is applied to a vbucket only the id is
propagated through the system - e.g. the "create-collection" system
marker item will only say the ID of the created collection and DCP only
sends the ID of created/deleted collections.

KV now assumes that if a client has enabled collections using HELLO that
the collection ID is encoded as a network-byte-order uint32 as the first
4-bytes of the key bytes.

Some KV objects now increase in size as the old DocNamespace was 1 byte
and now CollectionID stores uint32, however clients can reduce their
key sizes as there will no longer be a need to prefix keys with types.

With this commit all unit-tests pass, however

* DCP is broken, no clustering with collections enabled (MB-30397)
* unknown_command (setWithMeta etc...) with collections enabled won't
  work as the key/cid decoding must occur in KV (MB-30402)
* Further tidying is required, but will be done separately to reduce
  the size of this commit (MB-30521)

Change-Id: I583dd0e12b4b17057d325939a2da60ef16ff78b2
Reviewed-on: http://review.couchbase.org/96547
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>