MB-30302: Remove memory tracking around BucketLogger
Remove memory tracking in the BucketLogger so we track de-allocations
of our custom formatting correctly. We already track any allocations
made by spdlog while formatting so there is no point in not doing so.

Set the testing log level in the new BucketLogger to hide output in the
same way as before.

Change-Id: Iabca841e58b1f8598eccabe2054195a384313395
Reviewed-on: http://review.couchbase.org/96937
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>