MB-30397: [1/2] Simplify DCP open
Simplify DCP open so that only the HELLO collections flag
determines if collections can be supported by the consumer/producer.
DCP_OPEN_COLLECTIONS is no longer needed (no client should
be using it)

Updates the dcp open engine method so that collections are enabled
via the optional buffer, which can contain an empty buffer when
filtering is disabled.

Change-Id: Ie13dc8f488548e0a904305a157f93be94aa18163
Reviewed-on: http://review.couchbase.org/97064
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>