MB-30532: Improve the 'DCP Consumer snap-end mutation latency' perftest
CheckpointManager::checkAndAddNewCheckpoint is O(N),
where N = CheckpointManager::checkpointList.size().
This is the main cause of the regression in MB-30019, where the number
of checkpoints is reaches very high values (up to ~50000).

I want the perftest to catch better that scenario, so that we can
measure the optimization improvements coming with follow-up patches.
So, I update the test to prevent the ClosedUnrefCheckpointRemoverTask
from releasing checkpoints. Doing that, the number of checkpoints
increases constantly.

Again, I show before/after measurements with re-introducing the MB-30019
regression on my local environment.

[Before the regression]
=== DCP Consumer snapshot-end mutation - 20000 items === Latency (ns) =======================

                                Percentile
                         Median     95th     99th  Std Dev  Histogram of samples

Datatype::Raw            12.003   22.618   41.737    8.884  ▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▃▁▁▁▁▁▁▁▁▁▁▁▁▁
                                                            11             ns             22

[After the regression]
=== DCP Consumer snapshot-end mutation - 20000 items === Latency (ns) =======================

                                Percentile
                         Median     95th     99th  Std Dev  Histogram of samples

Datatype::Raw           287.665  628.839  752.913  182.681  ▁▁▄▅▄▆▇█▅▅▅▅▆▅▅▅▅▅▄▄▅▅▄▄▄▄▃▄▄▄▃▂
                                                            52             ns            628

Change-Id: I3ba0dbfccc417aa5a975afd2c6d1e861f2b368d6
Reviewed-on: http://review.couchbase.org/97726
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>