MB-30552: Move vb id formatting into class Vbid
Create a new class Vbid to initially handle the wrapping of
outputting vb id's in a consistent format, but to develop into
the main type of all vBucket usage.

The class 'Vbid' is located in memcached/vbucket.h in
preperation for future integration of vBucket ID's being
converted to this new type.

Change-Id: Iabc58c9a976bced1a77a1b7a6a77766c9c5019bd
Reviewed-on: http://review.couchbase.org/97996
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>