MB-30302: Preparation for change to runtime verbosity changing
In order to change logging verbosity within ep engine at runtime we
currently update a global log level atomic variable and consult it
when logging. With the introduction of spdlog in ep engine we can
instead change the logging level of each logger by using the spdlog
registry. In order to use the spldog registry we must pass it a
shared pointer to an spdlogger instance.

Change the loggers within ep engine to use shared pointers so that they
can be registered within the spdlog registry.

Change-Id: I27524e7961dbe7450739168a0d622ed22122acb7
Reviewed-on: http://review.couchbase.org/98309
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>